+++ idf_component_manager/cli/manifest.py
@@ -27,24 +27,51 @@ def schema():
     print_info(json.dumps(JSON_SCHEMA, indent=2))
 
 
-MANIFEST_COMPONENT_NAME_OPTION = [click.option('--component', default='main', help='Component name in the project')]
+MANIFEST_OPTION = [
+    click.option('--component', default='main', help='Component name in the project'),
+    click.option(
+        '-p',
+        '--path',
+        default=None,
+        help='Path to the component. The component name is ignored when the path is specified.')
+]
 
 
 @manifest.command()
-@add_options(PROJECT_DIR_OPTION + MANIFEST_COMPONENT_NAME_OPTION)
-def create(manager, component):
+@add_options(PROJECT_DIR_OPTION + MANIFEST_OPTION)
+def create(manager, component, path):
     """
     Create manifest file for the specified component.
+
+    By default:
+
+    If you run the command in the directory with project, the manifest
+    will be created in the "main" directory.
+
+    If you run the command in the directory with a component, the manifest
+    will be created right in that directory.
+
+    You can explicitly specify directory using the --path option.
     """
-    manager.create_manifest(component=component)
+    manager.create_manifest(component=component, path=path)
 
 
 @manifest.command()
-@add_options(PROJECT_DIR_OPTION + MANIFEST_COMPONENT_NAME_OPTION)
+@add_options(PROJECT_DIR_OPTION + MANIFEST_OPTION)
 @click.argument('dependency', required=True)
-def add_dependency(manager, component, dependency):
+def add_dependency(manager, component, path, dependency):
     """
-    Add dependency to the manifest file. For now we only support adding dependencies from the component registry.
+    Add dependency to the manifest file.
+
+    By default:
+
+    If you run the command in the directory with project, the dependency
+    will be added to the manifest in the "main" directory.
+
+    If you run the command in the directory with a component,
+    the dependency will be added to the manifest right in that directory.
+
+    You can explicitly specify directory using the --path option.
 
     \b
     Examples:
@@ -53,4 +80,4 @@ def add_dependency(manager, component, dependency):
     - $ compote manifest add-dependency example/cmp<=2.0.0
       would add component `example/cmp` with constraint `<=2.0.0`
     """
-    manager.add_dependency(dependency, component=component)
+    manager.add_dependency(dependency, component=component, path=path)